Move simulator harness to sim/tools
- move build.js logic back into build and remove 'full' option
- dev-tools/simulate logic moves into sim/tools
- sim/examples-random-player-ai.ts moved into sim/tools where its
used and sim/battle-stream-example deleted as sim/tools give plenty
of examples of its use.
