+++ sim/tools/exhaustive-runner.ts
@@ -1,239 +1,16 @@
 /**
- * Battle Simulator runners.
+ * Battle Simulator exhaustive runner.
  * Pokemon Showdown - http://pokemonshowdown.com/
  *
  * @license MIT
  */
 
-'use strict';
-
-import {ObjectReadWriteStream} from '../lib/streams';
-import * as BattleStreams from '../sim/battle-stream';
-import Dex = require('../sim/dex');
-import {RandomPlayerAI} from '../sim/examples/random-player-ai';
-import {PRNG, PRNGSeed} from '../sim/prng';
+import {ObjectReadWriteStream} from '../../lib/streams';
+import Dex = require('../dex');
+import {PRNG, PRNGSeed} from '../prng';
+import {RandomPlayerAI} from './random-player-ai';
+import {AIOptions, Runner} from './runner';
 const toId = Dex.getId;
-
-export interface AIOptions {
-	createAI: (stream: ObjectReadWriteStream<string>, options: AIOptions) => RandomPlayerAI;
-	move?: number;
-	mega?: number;
-	seed?: PRNG | PRNGSeed | null;
-	team?: PokemonSet[];
-}
-
-export interface RunnerOptions {
-	format: string;
-	prng?: PRNG | PRNGSeed | null;
-	p1options?: AIOptions;
-	p2options?: AIOptions;
-	input?: boolean;
-	output?: boolean;
-	error?: boolean;
-}
-
-export class Runner {
-	static readonly AI_OPTIONS: AIOptions = {
-		createAI: (s: ObjectReadWriteStream<string>, o: AIOptions) => new RandomPlayerAI(s, o),
-		move: 0.7,
-		mega: 0.6,
-	};
-
-	private readonly prng: PRNG;
-	private readonly p1options: AIOptions;
-	private readonly p2options: AIOptions;
-	private readonly format: string;
-	private readonly input: boolean;
-	private readonly output: boolean;
-	private readonly error: boolean;
-
-	constructor(options: RunnerOptions) {
-		this.format = options.format;
-
-		this.prng = (options.prng && !Array.isArray(options.prng)) ?
-			options.prng : new PRNG(options.prng);
-		this.p1options = Object.assign({}, Runner.AI_OPTIONS, options.p1options);
-		this.p2options = Object.assign({}, Runner.AI_OPTIONS, options.p2options);
-
-		this.input = !!options.input;
-		this.output = !!options.output;
-		this.error = !!options.error;
-	}
-
-	async run() {
-		const battleStream = new RawBattleStream(this.input);
-		const game = this.runGame(this.format, battleStream);
-		if (!this.error) return game;
-		return game.catch(err => {
-			console.log(`\n${battleStream.rawInputLog.join('\n')}\n`);
-			throw err;
-		});
-	}
-
-	private async runGame(format: string, battleStream: RawBattleStream) {
-		const streams = BattleStreams.getPlayerStreams(battleStream);
-		const spec = {formatid: format, seed: this.prng.seed};
-		const p1spec = this.getPlayerSpec("Bot 1", this.p1options);
-		const p2spec = this.getPlayerSpec("Bot 2", this.p2options);
-
-		const p1 = this.p1options.createAI(
-			streams.p1, Object.assign({seed: this.newSeed()}, this.p1options));
-		const p2 = this.p2options.createAI(
-			streams.p2, Object.assign({seed: this.newSeed()}, this.p2options));
-		// TODO: Use `await Promise.race([streams.omniscient.read(), p1, p2])` to avoid
-		// leaving these promises dangling once it no longer causes memory leaks (v8#9069).
-		/* tslint:disable:no-floating-promises */
-		p1.start();
-		p2.start();
-		/* tslint:enable:no-floating-promises */
-
-		streams.omniscient.write(`>start ${JSON.stringify(spec)}\n` +
-			`>player p1 ${JSON.stringify(p1spec)}\n` +
-			`>player p2 ${JSON.stringify(p2spec)}`);
-
-		let chunk;
-		// tslint:disable-next-line no-conditional-assignment
-		while ((chunk = await streams.omniscient.read())) {
-			if (this.output) console.log(chunk);
-		}
-		return streams.omniscient.end();
-	}
-
-	// Same as PRNG#generatedSeed, only deterministic.
-	// NOTE: advances this.prng's seed by 4.
-	private newSeed(): PRNGSeed {
-		return [
-			Math.floor(this.prng.next() * 0x10000),
-			Math.floor(this.prng.next() * 0x10000),
-			Math.floor(this.prng.next() * 0x10000),
-			Math.floor(this.prng.next() * 0x10000),
-		];
-	}
-
-	private getPlayerSpec(name: string, options: AIOptions) {
-		if (options.team) return {name, team: options.team};
-		return {name, seed: this.newSeed()};
-	}
-}
-
-class RawBattleStream extends BattleStreams.BattleStream {
-	readonly rawInputLog: string[];
-
-	private readonly input: boolean;
-
-	constructor(input: boolean) {
-		super();
-		this.input = !!input;
-		this.rawInputLog = [];
-	}
-
-	_write(message: string) {
-		if (this.input) console.log(message);
-		this.rawInputLog.push(message);
-		super._write(message);
-	}
-}
-
-// @ts-ignore
-export interface MultiRandomRunnerOptions extends RunnerOptions {
-	totalGames: number;
-	prng?: PRNG | PRNGSeed | null;
-	format?: string;
-	cycle?: boolean;
-	all?: boolean;
-	async?: boolean;
-}
-
-export class MultiRandomRunner {
-	static readonly FORMATS = [
-		'gen7randombattle', 'gen7randomdoublesbattle', 'gen7battlefactory', 'gen6randombattle', 'gen6battlefactory',
-		'gen5randombattle', 'gen4randombattle', 'gen3randombattle', 'gen2randombattle', 'gen1randombattle',
-	];
-
-	private readonly options: Partial<RunnerOptions>;
-	private readonly totalGames: number;
-	private readonly prng: PRNG;
-	private readonly format: string | undefined;
-	private readonly cycle: boolean;
-	private readonly all: boolean;
-	private readonly async: boolean;
-
-	private formatIndex: number;
-	private numGames: number;
-
-	constructor(options: MultiRandomRunnerOptions) {
-		this.options = Object.assign({}, options);
-
-		this.totalGames = options.totalGames;
-
-		this.prng = (options.prng && !Array.isArray(options.prng)) ?
-			options.prng : new PRNG(options.prng);
-		this.options.prng = this.prng;
-
-		this.format = options.format;
-		this.cycle = !!options.cycle;
-		this.all = !!options.all;
-
-		this.async = !!options.async;
-
-		this.formatIndex = 0;
-		this.numGames = 0;
-	}
-
-	async run() {
-		let games = [];
-		let format: string | false;
-		let lastFormat: string | false = false;
-		let failures = 0;
-		// tslint:disable-next-line no-conditional-assignment
-		while ((format = this.getNextFormat())) {
-			if (this.all && lastFormat && format !== lastFormat) {
-				if (this.async) await Promise.all(games);
-				games = [];
-			}
-
-			const seed = this.prng.seed;
-			const game = new Runner(Object.assign({format}, this.options)).run().catch(err => {
-				failures++;
-				console.error(
-					`Run \`node dev-tools/simulate multi 1 --format=${format} --seed=${seed.join()}\` ` +
-					`to debug (optionally with \`--output\` and/or \`--input\` for more info):\n`, err);
-			});
-
-			if (!this.async) await game;
-			games.push(game);
-			lastFormat = format;
-		}
-
-		if (this.async) await Promise.all(games);
-		return failures;
-	}
-
-	private getNextFormat() {
-		const FORMATS = MultiRandomRunner.FORMATS;
-		if (this.formatIndex > FORMATS.length) return false;
-
-		if (this.numGames++ < this.totalGames) {
-			if (this.format) {
-				return this.format;
-			} else if (this.all) {
-				return FORMATS[this.formatIndex];
-			} else if (this.cycle) {
-				const format = FORMATS[this.formatIndex];
-				this.formatIndex = (this.formatIndex + 1) % FORMATS.length;
-				return format;
-			} else {
-				return this.prng.sample(FORMATS);
-			}
-		} else if (this.all) {
-			this.numGames = 1;
-			this.formatIndex++;
-			return FORMATS[this.formatIndex];
-		}
-
-		return false;
-	}
-}
 
 interface Pools {
 	pokemon: Pool;
@@ -314,7 +91,7 @@ export class ExhaustiveRunner {
 			} catch (err) {
 				this.failures++;
 				console.error(
-					`\n\nRun \`node dev-tools/simulate exhaustive --cycles=${this.cycles} ` +
+					`\n\nRun \`node tools/simulate exhaustive --cycles=${this.cycles} ` +
 					`--format=${this.format} --seed${seed.join()}\`:\n`, err);
 			}
 		} while ((!this.maxGames || this.games < this.maxGames) &&
