+++ sim/tools/random-player-ai.ts
